feat: support trigger migration
