+++ packages/asset-packs/src/enums.ts
@@ -1,11 +1,11 @@
 import type { YGAlign, YGJustify } from '@dcl/ecs';
-import { AssetPacksComponentsVersions, getLatestComponentVersion } from './versioning';
+import { getLatestVersionName } from './versioning';
 
 export const ComponentName = {
   ACTION_TYPES: 'asset-packs::ActionTypes',
   ACTIONS: 'asset-packs::Actions',
-  COUNTER: getLatestComponentVersion(AssetPacksComponentsVersions.Counter).versionName,
-  TRIGGERS: 'asset-packs::Triggers',
+  COUNTER: getLatestVersionName('asset-packs::Counter'),
+  TRIGGERS: getLatestVersionName('asset-packs::Triggers'),
   STATES: 'asset-packs::States',
   COUNTER_BAR: 'asset-packs::CounterBar',
   ADMIN_TOOLS: 'asset-packs::AdminTools',
@@ -17,7 +17,6 @@ export const ComponentName = {
 } as const;
 
 export type ComponentName = (typeof ComponentName)[keyof typeof ComponentName];
-
 export enum TweenType {
   MOVE_ITEM = 'move_item',
   ROTATE_ITEM = 'rotate_item',
@@ -112,45 +111,8 @@ export enum ActionType {
   CALL_SCRIPT_METHOD = 'call_script_method',
 }
 
-export enum TriggerType {
-  /** @deprecated use ON_INPUT_ACTION instead */
-  ON_CLICK = 'on_click',
-  ON_INPUT_ACTION = 'on_input_action',
-  ON_STATE_CHANGE = 'on_state_change',
-  ON_SPAWN = 'on_spawn',
-  ON_TWEEN_END = 'on_tween_end',
-  ON_COUNTER_CHANGE = 'on_counter_change',
-  ON_PLAYER_ENTERS_AREA = 'on_player_enters_area',
-  ON_PLAYER_LEAVES_AREA = 'on_player_leaves_area',
-  ON_DELAY = 'on_delay',
-  ON_LOOP = 'on_loop',
-  ON_CLONE = 'on_clone',
-  ON_CLICK_IMAGE = 'on_click_image',
-  ON_DAMAGE = 'on_damage',
-  ON_GLOBAL_CLICK = 'on_global_click',
-  ON_GLOBAL_PRIMARY = 'on_global_primary',
-  ON_GLOBAL_SECONDARY = 'on_global_secondary',
-  ON_TICK = 'on_tick',
-  ON_HEAL_PLAYER = 'on_heal_player',
-  ON_PLAYER_SPAWN = 'on_player_spawn',
-}
-
-export enum TriggerConditionType {
-  WHEN_STATE_IS = 'when_state_is',
-  WHEN_STATE_IS_NOT = 'when_state_is_not',
-  WHEN_COUNTER_EQUALS = 'when_counter_equals',
-  WHEN_COUNTER_IS_GREATER_THAN = 'when_counter_is_greater_than',
-  WHEN_COUNTER_IS_LESS_THAN = 'when_counter_is_less_than',
-  WHEN_DISTANCE_TO_PLAYER_LESS_THAN = 'when_distance_to_player_less_than',
-  WHEN_DISTANCE_TO_PLAYER_GREATER_THAN = 'when_distance_to_player_greater_than',
-  WHEN_PREVIOUS_STATE_IS = 'when_previous_state_is',
-  WHEN_PREVIOUS_STATE_IS_NOT = 'when_previous_state_is_not',
-}
-
-export enum TriggerConditionOperation {
-  AND = 'and',
-  OR = 'or',
-}
+// Re-export trigger enums from versioning (source of truth)
+export { TriggerType, TriggerConditionType, TriggerConditionOperation } from './versioning';
 
 // Defined values instead of using from @dcl/ecs because Schemas doesn't support const enums
 export enum AlignMode {
