--- packages/asset-packs/src/versioning/components/index.ts
@@ -1,1 +1,0 @@
-export * from './counter';
