+++ packages/asset-packs/src/versioning/definitions/counter.ts
@@ -1,4 +1,5 @@
 import { Schemas } from '@dcl/ecs';
+import type { IEngine } from '@dcl/ecs';
 
 const COUNTER_BASE_NAME = 'asset-packs::Counter';
 
@@ -8,12 +9,16 @@ const CounterV0 = {
 };
 
 const CounterV1 = {
-  id: Schemas.Number,
-  value: Schemas.Int,
-  random: Schemas.Boolean,
+  ...CounterV0,
+  random: Schemas.Optional(Schemas.Boolean),
 };
 
 export const COUNTER_VERSIONS = [
   { versionName: COUNTER_BASE_NAME, component: CounterV0 },
   { versionName: `${COUNTER_BASE_NAME}-v1`, component: CounterV1 },
 ];
+
+export function defineCounterComponent(engine: IEngine) {
+  engine.defineComponent(COUNTER_BASE_NAME, CounterV0);
+  return engine.defineComponent(`${COUNTER_BASE_NAME}-v1`, CounterV1);
+}
