+++ packages/asset-packs/src/versioning/definitions/versionated-components.ts
