+++ packages/asset-packs/src/versioning/index.ts
@@ -1,2 +1,2 @@
 export * from './utils';
-export * from './components';
+export * from './definitions';
