+++ packages/asset-packs/src/versioning/utils.ts
@@ -1,84 +1,66 @@
 import type { ISchema, LastWriteWinElementSetComponentDefinition } from '@dcl/ecs';
 import type { IEngine } from '@dcl/ecs';
-import { COUNTER_VERSIONS } from './components';
+import {
+  COUNTER_VERSIONS,
+  TRIGGERS_VERSIONS,
+  defineCounterComponent,
+  defineTriggersComponent,
+} from './definitions';
 
 export type VersionedComponent = {
   versionName: string;
   component: Record<string, ISchema>;
 };
 
-export const AssetPacksComponentsVersions = {
-  Counter: COUNTER_VERSIONS,
-  // Actions: ACTIONS_VERSIONS,
-  // Triggers: TRIGGERS_VERSIONS,
-} as const;
+const VERSIONS_REGISTRY: Record<string, VersionedComponent[]> = {
+  'asset-packs::Counter': COUNTER_VERSIONS,
+  'asset-packs::Triggers': TRIGGERS_VERSIONS,
+};
 
-export function getLatestComponentVersion(versionedComponents: VersionedComponent[]) {
-  return versionedComponents[versionedComponents.length - 1];
-}
-
-function defineComponentVersions(engine: IEngine, versionedComponents: VersionedComponent[]) {
-  const components = versionedComponents.map(({ versionName, component }) => {
-    return engine.defineComponent(versionName, component);
-  });
-
-  return components;
-}
-
-export function getCompositeComponentVersion(
-  engine: IEngine,
-  versionedComponents: VersionedComponent[],
-) {
-  for (let i = versionedComponents.length - 1; i >= 0; i--) {
-    const { versionName } = versionedComponents[i];
-
-    const component = engine.getComponentOrNull(
-      versionName,
-    ) as LastWriteWinElementSetComponentDefinition<unknown> | null;
-
-    if (!component) continue;
-
-    const entities = [...engine.getEntitiesWith(component)];
-    if (entities.length === 0) continue;
-
-    return {
-      versionIndex: i,
-      versionName,
-      component,
-      entities,
-    };
-  }
-
-  return null;
-}
+export const getLatestVersionName = (baseName: string) => {
+  const versions = VERSIONS_REGISTRY[baseName];
+  return versions[versions.length - 1].versionName;
+};
 
 export function migrateVersionedComponent(
   engine: IEngine,
   versionedComponents: VersionedComponent[],
 ) {
-  const latestVersion = getLatestComponentVersion(versionedComponents);
-  const found = getCompositeComponentVersion(engine, versionedComponents);
+  const latestVersion = versionedComponents[versionedComponents.length - 1];
 
-  if (!found) return;
-  if (found.versionName === latestVersion.versionName) return;
+  for (let i = 0; i < versionedComponents.length - 1; i++) {
+    const oldVersion = versionedComponents[i];
 
-  const NewComponent = engine.getComponent(
-    latestVersion.versionName,
-  ) as LastWriteWinElementSetComponentDefinition<unknown>;
+    const oldComponent = engine.getComponentOrNull(
+      oldVersion.versionName,
+    ) as LastWriteWinElementSetComponentDefinition<unknown> | null;
 
-  for (const [entity, value] of found.entities) {
-    const oldValue = { ...value };
-    found.component.deleteFrom(entity);
-    NewComponent.createOrReplace(entity, oldValue);
+    if (!oldComponent) continue;
+
+    const entities = [...engine.getEntitiesWith(oldComponent)];
+    if (entities.length === 0) continue;
+
+    const newComponent = engine.getComponent(
+      latestVersion.versionName,
+    ) as LastWriteWinElementSetComponentDefinition<unknown>;
+
+    for (const [entity, value] of entities) {
+      oldComponent.deleteFrom(entity);
+      newComponent.createOrReplace(entity, { ...value });
+    }
+
+    engine.removeComponentDefinition(oldVersion.versionName);
   }
-
-  engine.removeComponentDefinition(found.versionName);
 }
 
 export function defineAssetPacksComponents(engine: IEngine) {
-  const Counter = defineComponentVersions(engine, AssetPacksComponentsVersions.Counter).pop()!;
-  // const Actions = defineComponentVersions(engine, AssetPacksComponentsVersions.Actions).pop()!;
-  // const Triggers = defineComponentVersions(engine, AssetPacksComponentsVersions.Triggers).pop()!;
+  const Counter = defineCounterComponent(engine);
+  const Triggers = defineTriggersComponent(engine);
 
-  return { Counter };
+  return { Counter, Triggers };
+}
+
+export function migrateAllAssetPacksComponents(engine: IEngine) {
+  migrateVersionedComponent(engine, COUNTER_VERSIONS);
+  migrateVersionedComponent(engine, TRIGGERS_VERSIONS);
 }
