+++ packages/inspector/src/lib/data-layer/host/composite-provider.ts
@@ -1,7 +1,7 @@
 import type { Scene } from '@dcl/schemas';
 import type { CompositeDefinition, Entity, IEngine } from '@dcl/ecs';
 import { Composite, CrdtMessageType, EntityMappingMode } from '@dcl/ecs';
-import { initComponents, migrateVersionedComponent, COUNTER_VERSIONS } from '@dcl/asset-packs';
+import { initComponents, migrateAllAssetPacksComponents } from '@dcl/asset-packs';
 import type { EditorComponents } from '../../sdk/components';
 import { EditorComponentNames } from '../../sdk/components';
 import type { FileSystemInterface } from '../types';
@@ -121,7 +121,7 @@ export class CompositeProvider implements StateProvider {
     fixNetworkEntityValues(this.engine);
     selectSceneEntity(this.engine);
     migrateSceneMetadata(this.engine);
-    migrateVersionedComponent(this.engine, COUNTER_VERSIONS);
+    migrateAllAssetPacksComponents(this.engine);
     createTagsComponent(this.engine);
   }
